prefabs
Created a prefab for every object so that way I do not have to spend a lot of time changing all of them.  I also added rigid body 2 to Ruby so that way she has some physics and next I need to add the Box Collider 2d to every object as a prefab probably.